+++ src/test/java/lotto/LotteryRetailerTest.java
@@ -7,6 +7,9 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.domain.LotteryRetailer;
+import lotto.domain.LottoNumberGenerator;
+import lotto.domain.LottoTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
